+++ Source/merchant/systems/Dnd5eCurrencyCalculator.ts
@@ -25,37 +25,32 @@ export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
         // Create a Consumable spell scroll on the Inventory tab
         if ( (itemData.type === "spell")) {
             const scroll = await this.createScroll(itemData);
-            itemData = scroll.data;
+			// @ts-ignore
+			return caller.callSuperOnDropItemCreate(scroll);
         }
 
         return caller.callSuperOnDropItemCreate(itemData);
     }
 
-	async createScrollFromSpell(spell: any) {
-
-		// Get spell data
-		const itemData =  spell.toObject();
-		const {actionType, description, source, activation, duration, target, range, damage, save, level} = itemData.data;
+	async createScrollFromSpell(spell: PropertiesToSource<ItemData>): Promise<PropertiesToSource<ItemData>> {
+		const itemData =  spell;
+		// const {actionType, description, source, activation, duration, target, range, damage, save, level} = itemData.data;
+		// @ts-ignore
+		let level = spell.data.level;
+		// @ts-ignore
+		let description = spell.data.description;
+		// @ts-ignore
 
 		// Get scroll data
 		// @ts-ignore
 		const scrollUuid = `Compendium.${CONFIG.DND5E.sourcePacks.ITEMS}.${CONFIG.DND5E.spellScrollIds[level]}`;
-		const scrollItem = fromUuid(scrollUuid);
-		if (scrollItem === undefined) {
-			return undefined;
-		}
+		const scrollItem = await fromUuid(scrollUuid);
 		// @ts-ignore
-		const scrollData = scrollItem.data;
-		if (scrollData === undefined) {
-			return undefined;
-		}
-
+		const scrollData = scrollItem.toObject();
+		delete scrollData._id;
 		// Split the scroll description into an intro paragraph and the remaining details
-		let scrollDescription = '';
-		if (scrollData !== undefined && scrollData.data !== undefined && scrollData.data.description !== undefined) {
-			scrollDescription = scrollData.data.description.value;
-		}
-		const pdel = '</p>';
+		const scrollDescription = scrollData.data.description.value;
+		const pdel = "</p>";
 		const scrollIntroEnd = scrollDescription.indexOf(pdel);
 		const scrollIntro = scrollDescription.slice(0, scrollIntroEnd + pdel.length);
 		const scrollDetails = scrollDescription.slice(scrollIntroEnd + pdel.length);
@@ -63,29 +58,18 @@ export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
 		// Create a composite description from the scroll description and the spell details
 		const desc = `${scrollIntro}<hr/><h3>${itemData.name} (Level ${level})</h3><hr/>${description.value}<hr/><h3>Scroll Details</h3><hr/>${scrollDetails}`;
 
-		// Create the spell scroll data
-		const spellScrollData = foundry.utils.mergeObject(scrollData, {
-			name: `${(<Game>game).i18n.localize("DND5E.SpellScroll")}: ${itemData.name}`,
-			img: itemData.img,
-			data: {
-				"description.value": desc.trim(),
-				source,
-				actionType,
-				activation,
-				duration,
-				target,
-				range,
-				damage,
-				save,
-				level
-			}
-		});
+		let clone = duplicate(itemData);
+		clone.name = `${(<Game>game).i18n.localize("DND5E.SpellScroll")}: ${itemData.name}`;
+		clone.img = scrollData.img
+		clone.type = "consumable";
 		// @ts-ignore
-		return new this(spellScrollData);
+		clone.data.description.value = desc.trim()
+		// @ts-ignore
+		return clone;
 	}
 
-	async createScroll(itemData: PropertiesToSource<ItemData>) {
-        return await this.createScrollFromSpell(itemData);
+	async createScroll(itemData: PropertiesToSource<ItemData>): Promise<PropertiesToSource<ItemData>> {
+        return this.createScrollFromSpell(itemData);
     }
 
     actorCurrency(actor: Actor) {
@@ -108,13 +92,11 @@ export default class Dnd5eCurrencyCalculator extends CurrencyCalculator {
         buyerFundsAsPlatinum += buyerFunds["sp"] / conversionRates["gp"] / conversionRates["ep"] / conversionRates["sp"];
         buyerFundsAsPlatinum += buyerFunds["cp"] / conversionRates["gp"] / conversionRates["ep"] / conversionRates["sp"] / conversionRates["cp"];
 
-        console.log(`buyerFundsAsPlatinum : ${buyerFundsAsPlatinum}`);
 
         return buyerFundsAsPlatinum;
     }
 
     updateActorWithNewFunds(buyer: Actor, buyerFunds: any) {
-        console.log("Merchant sheet | buyer and funds", buyer,buyerFunds)
         buyer.update({ "data.currency": buyerFunds });
     }