+++ Source/merchant/systems/Wfrp4eCurrencyCalculator.ts
@@ -16,71 +16,9 @@ import Logger from "../../Utils/Logger";
 export default class Wfrp4eCurrencyCalculator extends CurrencyCalculator {
 
     async onDropItemCreate(itemData: PropertiesToSource<ItemData>, caller: MerchantSheet) {
-        // Create a Consumable spell scroll on the Inventory tab
-        if ( (itemData.type === "spell")) {
-            const scroll = await this.createScroll(itemData);
-            itemData = scroll.data;
-        }
-
         return caller.callSuperOnDropItemCreate(itemData);
     }
 
-	async createScrollFromSpell(spell: any) {
-
-		// Get spell data
-		const itemData =  spell.toObject();
-		const {actionType, description, source, activation, duration, target, range, damage, save, level} = itemData.data;
-
-		// Get scroll data
-		// @ts-ignore
-		const scrollUuid = `Compendium.${CONFIG.DND5E.sourcePacks.ITEMS}.${CONFIG.DND5E.spellScrollIds[level]}`;
-		const scrollItem = fromUuid(scrollUuid);
-		if (scrollItem === undefined) {
-			return undefined;
-		}
-		// @ts-ignore
-		const scrollData = scrollItem.data;
-		if (scrollData === undefined) {
-			return undefined;
-		}
-
-		// Split the scroll description into an intro paragraph and the remaining details
-		let scrollDescription = '';
-		if (scrollData !== undefined && scrollData.data !== undefined && scrollData.data.description !== undefined) {
-			scrollDescription = scrollData.data.description.value;
-		}
-		const pdel = '</p>';
-		const scrollIntroEnd = scrollDescription.indexOf(pdel);
-		const scrollIntro = scrollDescription.slice(0, scrollIntroEnd + pdel.length);
-		const scrollDetails = scrollDescription.slice(scrollIntroEnd + pdel.length);
-
-		// Create a composite description from the scroll description and the spell details
-		const desc = `${scrollIntro}<hr/><h3>${itemData.name} (Level ${level})</h3><hr/>${description.value}<hr/><h3>Scroll Details</h3><hr/>${scrollDetails}`;
-
-		// Create the spell scroll data
-		const spellScrollData = foundry.utils.mergeObject(scrollData, {
-			name: `${(<Game>game).i18n.localize("DND5E.SpellScroll")}: ${itemData.name}`,
-			img: itemData.img,
-			data: {
-				"description.value": desc.trim(),
-				source,
-				actionType,
-				activation,
-				duration,
-				target,
-				range,
-				damage,
-				save,
-				level
-			}
-		});
-		// @ts-ignore
-		return new this(spellScrollData);
-	}
-
-	async createScroll(itemData: PropertiesToSource<ItemData>) {
-        return await this.createScrollFromSpell(itemData);
-    }
 
     actorCurrency(actor: Actor) {
         // @ts-ignore